separate config and datanode impl
